backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv
pydantic>=2.0
google-generativeai
pytest
requests
httpx
